Put all auth functions in auth module
