+++ source/funwithflags/use_cases/__init__.py
@@ -1,8 +1,2 @@
 """Module for use cases."""
-from .fresh_login import fresh_login
-from .register import register
-from .login import login
-from .logout import logout
-from .read_user import read_user_basic
-from .refresh_access_token import refresh_access_token
-from .update_user import update_user
+from .auth import fresh_login, login, logout, read_user_basic, refresh_access_token, register, update_user
