+++ source/integration_tests/test_integration_03_register.py
@@ -1,5 +1,4 @@
 """Integration test for signup."""
-import json
 import pytest
 
 from funwithflags.definitions import RegisterRequest
